test: factor shared Qwen3 XML/Coder regression tests into one file
Both the XML and Coder tool parsers were tested against nearly
identical regression scenarios in their respective files (string
"null" preservation, anyOf nullable schemas, double-encoded objects,
content with literal XML structural tags, content with param-like
lines, etc.).  Split the shared expectations into a single file with
a parametrized parser fixture so that:

* the same intent is tested against BOTH parsers automatically;
* divergent behaviour is caught immediately instead of drifting;
* parser-specific quirks (XML SAX double-close brace, char-by-char
  boolean streaming, Coder speculative-decoding chunk loss, etc.)
  stay in their parser-specific test file.

New: tests/tool_parsers/test_qwen3_xml_coder_shared.py exposes a
``parser_cls`` fixture parametrized over Qwen3XMLToolParser and
Qwen3CoderToolParser.  Each shared test runs twice and prints
``[xml]``/``[coder]`` in the test id.

Removed duplicates from:
* tests/tool_parsers/test_qwen3xml_tool_parser.py: anyOf object
  param (streaming + non-streaming), string null preservation, anyOf
  integer/null type detection, content with structural tags
  (streaming + non-streaming), content with param-like lines
  (streaming + non-streaming), double-encoded object (streaming +
  non-streaming).
* tests/tool_parsers/test_qwen3coder_tool_parser.py: anyOf parameter
  not double encoded, string null preservation, anyOf string/null
  numeric value, content with XML structural tags (streaming +
  non-streaming), content with param-like lines (streaming +
  non-streaming), double-encoded object (streaming + non-streaming),
  content param with tool_call tag (streaming + non-streaming —
  redundant with content_with_xml_structural_tags).

Removed: tests/tool_parsers/test_qwen36_bugs.py.  Its two scenarios
(XML array containing JSON ``true``, Coder two complete tool calls
in a single streaming delta) are now in the shared file as
``test_array_with_json_bool`` and
``test_two_tool_calls_in_one_streaming_chunk``, both running against
both parsers.

Net effect: 209 -> 183 tests, 0 failures, identical coverage.
