+++ tests/tool_parsers/test_qwen3xml_tool_parser.py
@@ -172,93 +172,6 @@ class TestQwen3xmlToolParser(ToolParserTests):
         assert "I hope this helps!" in all_content, "Free text after the last tool call should be emitted."
 
 
-def test_qwen36_anyof_parameter_xml_not_double_encoded(qwen3_tokenizer):
-    tools = [
-        ChatCompletionToolsParam(
-            type="function",
-            function={
-                "name": "update_record",
-                "parameters": {
-                    "type": "object",
-                    "properties": {
-                        # anyOf schema — no top-level "type" key
-                        "data": {
-                            "anyOf": [{"type": "object"}, {"type": "null"}],
-                        },
-                    },
-                },
-            },
-        )
-    ]
-
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=tools)
-    model_output = (
-        "<tool_call>\n"
-        "<function=update_record>\n"
-        '<parameter=data>{"key": "value", "count": 42}</parameter>\n'
-        "</function>\n"
-        "</tool_call>"
-    )
-    request = ChatCompletionRequest(model=MODEL, messages=[], tools=tools)
-    result = parser.extract_tool_calls(model_output, request=request)
-
-    assert result.tools_called
-    assert len(result.tool_calls) == 1
-    args = json.loads(result.tool_calls[0].function.arguments)
-
-    assert isinstance(args["data"], dict), (
-        f"anyOf parameter was double-encoded: data={args['data']!r}. "
-        "StreamingXMLToolCallParser._get_param_type ignores anyOf schemas."
-    )
-    assert args["data"] == {"key": "value", "count": 42}
-
-
-def test_qwen36_anyof_parameter_xml_streaming_not_double_encoded(qwen3_tokenizer):
-  
-    tools = [
-        ChatCompletionToolsParam(
-            type="function",
-            function={
-                "name": "update_record",
-                "parameters": {
-                    "type": "object",
-                    "properties": {
-                        "data": {
-                            "anyOf": [{"type": "object"}, {"type": "null"}],
-                        },
-                    },
-                },
-            },
-        )
-    ]
-
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=tools)
-    request = ChatCompletionRequest(model=MODEL, messages=[], tools=tools)
-
-    # Deltas are pre-formed XML element chunks (one element per delta),
-    # which is the same pattern used by speculative decoding.
-    deltas = [
-        "<tool_call>",
-        "\n<function=update_record>",
-        '\n<parameter=data>{"key": "value", "count": 42}</parameter>',
-        "\n</function>",
-        "\n</tool_call>",
-    ]
-
-    reconstructor = run_tool_extraction_streaming(
-        parser,
-        deltas,
-        request,
-        assert_one_tool_per_delta=False,
-    )
-
-    assert len(reconstructor.tool_calls) == 1
-    args = json.loads(reconstructor.tool_calls[0].function.arguments)
-    assert isinstance(args["data"], dict), (
-        f"anyOf parameter was double-encoded in streaming: data={args['data']!r}"
-    )
-
-
 def test_qwen36_xml_streaming_double_close_brace(qwen3_tokenizer):
     tools = [
         ChatCompletionToolsParam(
@@ -360,58 +273,8 @@ def test_xml_streaming_parallel_tool_calls_preformed_chunks(qwen3_tokenizer):
 
 
 # ---------------------------------------------------------------------------
-# Bug-confirmation tests (regressions to FIX)
+# XML-specific streaming bugs (Coder parser is not affected)
 # ---------------------------------------------------------------------------
-
-
-def test_xml_string_null_value_not_emptied(qwen3_tokenizer):
-    """
-    Bug A: _convert_param_value intercepts "null" before the type check.
-    For a STRING parameter with value "null", the parser should output
-    the JSON string "null", not an empty string "".
-
-    Root cause: `if param_value.lower() == "null": return None` runs first,
-    then _convert_for_json_streaming(None, "string") returns "", so the
-    closing-quote _end_element emits "" instead of "null".
-    """
-    tools = [
-        ChatCompletionToolsParam(
-            type="function",
-            function={
-                "name": "search",
-                "parameters": {
-                    "type": "object",
-                    "properties": {
-                        "query": {"type": "string"},
-                    },
-                },
-            },
-        )
-    ]
-
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=tools)
-    model_output = (
-        "<tool_call>\n"
-        "<function=search>\n"
-        "<parameter=query>null</parameter>\n"
-        "</function>\n"
-        "</tool_call>"
-    )
-    request = ChatCompletionRequest(model=MODEL, messages=[], tools=tools)
-    result = parser.extract_tool_calls(model_output, request=request)
-
-    assert result.tools_called
-    assert len(result.tool_calls) == 1
-    args = json.loads(result.tool_calls[0].function.arguments)
-
-    assert "query" in args, f"Parameter 'query' missing from args: {args!r}"
-    assert args["query"] == "null", (
-        f"String parameter with literal value 'null' was incorrectly converted. "
-        f"Got: {args['query']!r}. "
-        f"Expected: 'null' (the string). "
-        f"_convert_param_value returns None before checking type, "
-        f"then _convert_for_json_streaming(None, 'string') returns ''."
-    )
 
 
 def test_xml_streaming_boolean_true_not_false(qwen3_tokenizer):
@@ -534,360 +397,3 @@ def test_xml_streaming_string_null_last_char_not_dropped(qwen3_tokenizer):
         f"and _convert_for_json_streaming(None, 'string') returns '', "
         f"so the final delta is empty and the 'l' is never emitted."
     )
-
-
-def test_xml_anyof_integer_null_type_detected(qwen3_tokenizer):
-    """
-    Bug C: _get_param_type only returns non-string for anyOf schemas that
-    contain "object" or "array". For anyOf: [{type: "integer"}, {type: "null"}]
-    it falls through and returns "string", so integer parameters with
-    nullable schemas are incorrectly quoted and not converted.
-    """
-    tools = [
-        ChatCompletionToolsParam(
-            type="function",
-            function={
-                "name": "set_count",
-                "parameters": {
-                    "type": "object",
-                    "properties": {
-                        "count": {
-                            "anyOf": [{"type": "integer"}, {"type": "null"}],
-                        },
-                    },
-                },
-            },
-        )
-    ]
-
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=tools)
-    model_output = (
-        "<tool_call>\n"
-        "<function=set_count>\n"
-        "<parameter=count>42</parameter>\n"
-        "</function>\n"
-        "</tool_call>"
-    )
-    request = ChatCompletionRequest(model=MODEL, messages=[], tools=tools)
-    result = parser.extract_tool_calls(model_output, request=request)
-
-    assert result.tools_called
-    assert len(result.tool_calls) == 1
-    args = json.loads(result.tool_calls[0].function.arguments)
-
-    assert args["count"] == 42, (
-        f"anyOf integer+null: expected int 42, got {args['count']!r}. "
-        f"_get_param_type only checks for object/array in anyOf schemas, "
-        f"so integer anyOf schemas fall back to 'string', causing '42' "
-        f"to be returned as the JSON string '\"42\"' instead of the number 42."
-    )
-
-
-# ---------------------------------------------------------------------------
-# Regression: XML structural tags as literal text inside string parameters
-# ---------------------------------------------------------------------------
-
-_WRITE_FILE_TOOLS_XML = [
-    ChatCompletionToolsParam(
-        type="function",
-        function={
-            "name": "write_file",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "file_path": {"type": "string"},
-                    "content": {"type": "string"},
-                },
-            },
-        },
-    )
-]
-
-# Python content that contains all four XML structural tags as literal strings.
-# When qwen3xml encounters "</parameter>" inside the content value it
-# currently treats it as the structural end of the <parameter=content> element,
-# truncating the value and creating a spurious "query" parameter from the text
-# that follows the fake </parameter>.
-_XML_TAGS_IN_CONTENT_XML = (
-    'char_deltas = [\n'
-    '    "<tool_call>\\n",\n'
-    '    "<parameter=query>\\n",\n'
-    '    "\\n</parameter>\\n",\n'
-    '    "</function>\\n",\n'
-    ']\n'
-)
-
-_WRITE_FILE_XML_TAGS_OUTPUT_XML = (
-    "<tool_call>\n"
-    "<function=write_file>\n"
-    "<parameter=file_path>\ntest.py\n</parameter>\n"
-    f"<parameter=content>\n{_XML_TAGS_IN_CONTENT_XML}</parameter>\n"
-    "</function>\n"
-    "</tool_call>\n"
-)
-
-
-def test_xml_streaming_content_with_structural_xml_tags(qwen3_tokenizer):
-    """Streaming variant: pre-formed chunks, full content in one delta."""
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=_WRITE_FILE_TOOLS_XML)
-    request = ChatCompletionRequest(
-        model=MODEL, messages=[], tools=_WRITE_FILE_TOOLS_XML
-    )
-
-    char_deltas = [
-        "<tool_call>\n",
-        "<function=write_file>\n",
-        "<parameter=file_path>\ntest.py\n</parameter>\n",
-        f"<parameter=content>\n{_XML_TAGS_IN_CONTENT_XML}</parameter>\n",
-        "</function>\n",
-        "</tool_call>\n",
-    ]
-
-    reconstructor = run_tool_extraction_streaming(
-        parser,
-        char_deltas,
-        request,
-        assert_one_tool_per_delta=False,
-    )
-
-    assert len(reconstructor.tool_calls) == 1
-    assert reconstructor.tool_calls[0].function.name == "write_file"
-
-    args = json.loads(reconstructor.tool_calls[0].function.arguments)
-
-    assert list(args.keys()) == ["file_path", "content"], (
-        f"Unexpected parameter keys (spurious params from embedded tags?): "
-        f"{list(args.keys())}"
-    )
-    assert args["file_path"] == "test.py"
-    expected_content = _XML_TAGS_IN_CONTENT_XML.rstrip("\n")
-    assert args["content"] == expected_content, (
-        f"content was truncated or corrupted by embedded XML tags.\n"
-        f"Got:      {args.get('content')!r}\n"
-        f"Expected: {expected_content!r}"
-    )
-
-
-def test_xml_nonstreaming_content_with_structural_xml_tags(qwen3_tokenizer):
-    """Regression: string parameter containing </parameter>, </function>,
-    </tool_call> as literal text must be extracted intact.
-
-    Bug: the SAX pre-processor (_preprocess_xml_chunk) returns
-    ``safe_text + "</parameter>"`` when it sees ``</parameter>`` inside the
-    accumulated parameter buffer, terminating the current parameter too early.
-    The text that follows the spurious closing tag is then misinterpreted as a
-    new parameter named "query", creating a ghost parameter and truncating
-    the real "content" value.
-
-    Expected: exactly two parameters -- file_path and content -- with content
-    equal to the full Python snippet including the embedded XML tags.
-    """
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=_WRITE_FILE_TOOLS_XML)
-    request = ChatCompletionRequest(
-        model=MODEL, messages=[], tools=_WRITE_FILE_TOOLS_XML
-    )
-    result = parser.extract_tool_calls(_WRITE_FILE_XML_TAGS_OUTPUT_XML, request=request)
-
-    assert result.tools_called
-    assert len(result.tool_calls) == 1
-    assert result.tool_calls[0].function.name == "write_file"
-
-    args = json.loads(result.tool_calls[0].function.arguments)
-
-    assert list(args.keys()) == ["file_path", "content"], (
-        f"Unexpected parameter keys (spurious params created from embedded tags?): "
-        f"{list(args.keys())}. "
-        f"_preprocess_xml_chunk sees '</parameter>' inside the accumulated "
-        f"_pre_param_buffer and terminates the parameter early; the text after "
-        f"'<parameter=query>' becomes a ghost 'query' parameter."
-    )
-    assert args["file_path"] == "test.py"
-    expected_content = _XML_TAGS_IN_CONTENT_XML.rstrip("\n")
-    assert args["content"] == expected_content, (
-        f"content was truncated or corrupted by embedded XML tags. "
-        f"Got:      {args.get('content')!r}\n"
-        f"Expected: {expected_content!r}"
-    )
-
-
-# File content whose lines ARE standalone </parameter> and <parameter=NAME>
-# tokens (preceded by \n). This simulates writing a Jinja2 template, a test
-# fixture for the parser, or any file that references the tool-call format.
-# "new_string" is intentionally NOT a parameter of write_file (schema has
-# "file_path" and "content"), so the schema filter must prevent it from being
-# treated as a structural boundary.
-_CONTENT_WITH_PARAM_LIKE_LINES_XML = (
-    'TOOL_CALL_TEMPLATE = """\n'
-    "</parameter>\n"
-    "<parameter=new_string>\n"
-    "#!/usr/bin/env python3\n"
-    "</parameter>\n"
-    '"""\n'
-)
-
-_WRITE_FILE_PARAM_LIKE_LINES_OUTPUT_XML = (
-    "<tool_call>\n"
-    "<function=write_file>\n"
-    "<parameter=file_path>\ntest_template.py\n</parameter>\n"
-    f"<parameter=content>\n{_CONTENT_WITH_PARAM_LIKE_LINES_XML}</parameter>\n"
-    "</function>\n"
-    "</tool_call>\n"
-)
-
-
-def test_xml_nonstreaming_content_with_param_like_lines(qwen3_tokenizer):
-    """Non-streaming: file content containing </parameter> and <parameter=NAME>
-    on their own lines must not be truncated at the first </parameter> or
-    create spurious extra parameters.  Requires schema-based filtering so that
-    "new_string" (not a real parameter of write_file) is ignored.
-    """
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=_WRITE_FILE_TOOLS_XML)
-    request = ChatCompletionRequest(
-        model=MODEL, messages=[], tools=_WRITE_FILE_TOOLS_XML
-    )
-    result = parser.extract_tool_calls(_WRITE_FILE_PARAM_LIKE_LINES_OUTPUT_XML, request=request)
-
-    assert result.tools_called
-    assert len(result.tool_calls) == 1
-    assert result.tool_calls[0].function.name == "write_file"
-
-    args = json.loads(result.tool_calls[0].function.arguments)
-    assert list(args.keys()) == ["file_path", "content"], (
-        f"Spurious parameters created: {list(args.keys())}"
-    )
-    assert args["file_path"] == "test_template.py"
-    expected = _CONTENT_WITH_PARAM_LIKE_LINES_XML.rstrip("\n")
-    assert args["content"] == expected, (
-        f"content truncated or wrong: {args.get('content')!r}"
-    )
-
-
-def test_xml_streaming_content_with_param_like_lines(qwen3_tokenizer):
-    """Streaming: file content containing </parameter> and <parameter=NAME> on
-    their own lines — split into one chunk per structural token — must not
-    cause spurious extra parameters.
-
-    The critical scenario: chunk 5 is '</parameter>\\n' arriving ALONE so
-    the streaming buffer has nothing after it (rest='') which previously
-    triggered the 'not rest → structural' fallback, ending the 'content'
-    parameter prematurely.  After the schema fix, the subsequent
-    '<parameter=new_string>' is recognised as non-structural and the full
-    content is preserved.
-    """
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=_WRITE_FILE_TOOLS_XML)
-    request = ChatCompletionRequest(
-        model=MODEL, messages=[], tools=_WRITE_FILE_TOOLS_XML
-    )
-
-    char_deltas = [
-        "<tool_call>\n",
-        "<function=write_file>\n",
-        "<parameter=file_path>\ntest_template.py\n</parameter>\n",
-        '<parameter=content>\nTOOL_CALL_TEMPLATE = """\n',
-        "</parameter>\n",               # first literal close — alone in its delta
-        "<parameter=new_string>\n",     # literal new-param line
-        "#!/usr/bin/env python3\n",
-        "</parameter>\n",               # second literal close
-        '"""\n',
-        "</parameter>\n",               # REAL close of content
-        "</function>\n",
-        "</tool_call>\n",
-    ]
-
-    reconstructor = run_tool_extraction_streaming(
-        parser,
-        char_deltas,
-        request,
-        assert_one_tool_per_delta=False,
-    )
-
-    assert len(reconstructor.tool_calls) == 1, (
-        f"Expected 1 tool call, got {len(reconstructor.tool_calls)}: "
-        f"{[tc.function.name for tc in reconstructor.tool_calls]}"
-    )
-    assert reconstructor.tool_calls[0].function.name == "write_file"
-    args = json.loads(reconstructor.tool_calls[0].function.arguments)
-    assert list(args.keys()) == ["file_path", "content"], (
-        f"Spurious parameters created: {list(args.keys())}"
-    )
-    assert args["file_path"] == "test_template.py"
-    expected = _CONTENT_WITH_PARAM_LIKE_LINES_XML.rstrip("\n")
-    assert args["content"] == expected, (
-        f"content truncated or wrong: {args.get('content')!r}"
-    )
-
-
-_OBJECT_PARAM_TOOLS_XML = [
-    ChatCompletionToolsParam(
-        type="function",
-        function={
-            "name": "process",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "name": {"type": "string"},
-                    "data": {"type": "object"},
-                },
-            },
-        },
-    )
-]
-
-_DOUBLE_ENCODED_OBJECT_OUTPUT_XML = (
-    "<tool_call>\n"
-    "<function=process>\n"
-    "<parameter=name>\nhello\n</parameter>\n"
-    "<parameter=data>\n\"{'key': 'value', 'n': 1}\"\n</parameter>\n"
-    "</function>\n"
-    "</tool_call>\n"
-)
-
-
-def test_xml_nonstreaming_double_encoded_object_param(qwen3_tokenizer):
-    """Non-streaming: model trained with buggy template (json.dumps(str(dict)))
-    outputs object args as a JSON-encoded Python repr.  Parser must recover
-    the real dict via double-decode.
-    """
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=_OBJECT_PARAM_TOOLS_XML)
-    request = ChatCompletionRequest(
-        model=MODEL, messages=[], tools=_OBJECT_PARAM_TOOLS_XML
-    )
-    result = parser.extract_tool_calls(
-        _DOUBLE_ENCODED_OBJECT_OUTPUT_XML, request=request
-    )
-
-    assert result.tools_called
-    assert len(result.tool_calls) == 1
-    args = json.loads(result.tool_calls[0].function.arguments)
-    assert args["name"] == "hello"
-    assert isinstance(args["data"], dict), (
-        f"Expected dict, got {type(args['data'])}: {args['data']!r}"
-    )
-    assert args["data"] == {"key": "value", "n": 1}
-
-
-def test_xml_streaming_double_encoded_object_param(qwen3_tokenizer):
-    """Streaming: same double-encoded object parameter scenario."""
-    parser = Qwen3XMLToolParser(qwen3_tokenizer, tools=_OBJECT_PARAM_TOOLS_XML)
-    request = ChatCompletionRequest(
-        model=MODEL, messages=[], tools=_OBJECT_PARAM_TOOLS_XML
-    )
-    reconstructor = run_tool_extraction_streaming(
-        parser,
-        _DOUBLE_ENCODED_OBJECT_OUTPUT_XML,
-        request,
-        assert_one_tool_per_delta=False,
-    )
-    assert len(reconstructor.tool_calls) == 1
-    args = json.loads(reconstructor.tool_calls[0].function.arguments)
-    assert args["name"] == "hello"
-    assert isinstance(args["data"], dict), (
-        f"Expected dict, got {type(args['data'])}: {args['data']!r}"
-    )
-    assert args["data"] == {"key": "value", "n": 1}
-
-
-# ============================================================================
-# Qwen 3.6 Bug Confirmations (placeholder, truncated test removed)
-# ============================================================================
-
